refactor(cleanup): 사소한 패키지 구조 변경 및 문서 정리
