+++ DECISIONS.md
@@ -55,28 +55,28 @@
   - 결정: List로 변경한다.
   - 이유: MVP 개발 이후 책임 분배에 따라 필요없어짐.
 
-- [ ] 2025-11-02: `Lotto`의 생성자 스코프를 바꿔도 될까??
+- [x] 2025-11-02: `Lotto`의 생성자 스코프를 바꿔도 될까??
   - 고민: `LottoTest`는 위치만 옮기지 않으면 내부 코드 수정이 가능한 것 같다.
     - 프로그래밍 요구사항 1: "프로그래밍 요구 사항에서 달리 명시하지 않는 한 파일, 패키지 등의 이름을 바꾸거나 이동하지 않는다."
     - 프로그래밍 요구사항 3: "`Lotto`의 패키지를 변경할 수 있다."
   - 결정: 변경하지 않는다.
-  - 이유: 
+  - 이유: LottoTest 내부 코드 수정이 가능하지만 모험을 하기 싫다. 제공된 테스트는 원본을 남긴다.
 
-- [ ] 2025-11-02: `Lotto`의 로또 번호 범위 검증
+- [x] 2025-11-02: `Lotto`의 로또 번호 범위 검증
   - 고민: `Lotto`가 스스로 (1~45) 범위를 검사하는게 자연스럽다.
     - 하지만, 보너스 번호를 생성할 때는 `Lotto`를 재활용 하기 어렵다.
     - 그래서 `LottoNumberToken`을 뒀는데, 겹치는 검증이 많고, 클래스가 길어진다.
     - 공용 검증 로직을 도입해야하나? 
-  - 결정: 
-  - 이유: 
+  - 결정: 검증 보다는 특정 비즈니스 규칙을 담지 않는 조건들에 대한 조건을 util 에서 반환
+  - 이유: 캡슐화된 예외를 챙기면서 클래스 볼륨을 줄인다.
 
 ### answer
 
-- [ ] 2025-11-01: 당첨 정답(`answer`) 도메인의 예외 매서드 볼륨과 캡슐화 사이의 고민
+- [x] 2025-11-01: 당첨 정답(`answer`) 도메인의 예외 매서드 볼륨과 캡슐화 사이의 고민
   - 고민: 입력에 대한 세분화된 예외 메시지를 던지고 싶어서 예외 메서드를 좀 많게 구성했는데,
     - 생각보다 각 클래스(`WinningCondition`, `LottoNumberToken`)의 볼륨이 커졌다.
     - 아니면 각 관심사에 맞는 예외를 던지게끔 캡슐화를 유지할까?
-      - 이 경우 (`LottoNumberTokens`) 같은 중간 클래스를 하나 둬야 할 것 같긴 하다.
+    - 이 경우 (`LottoNumberTokens`) 같은 중간 클래스를 하나 둬야 할 것 같긴 하다.
     - 아니면 공통으로 사용되는 예외(빈 입력, 숫자만 포함되는지)는 public으로 열어두게끔 리팩토링 할까?
   - 결정: 일단은 그대로 진행해보되, 시간이 남는다면 고려해본다.
