+++ README.md
@@ -107,5 +107,5 @@
 ---
 
 ### 💭 기타
-- [ ] 결정 내리지 못한 고민들 결정하기
-- [ ] README에 예외 상황 정리하기
+- [x] 결정 내리지 못한 고민들 결정하기
+- [x] README에 예외 상황 정리하기
