+++ src/main/java/lotto/controller/LottoController.java
@@ -4,9 +4,9 @@ import java.util.List;
 import lotto.application.dto.ProfitRateDto;
 import lotto.application.dto.PurchasedLottoDto;
 import lotto.application.dto.WinningStatisticDto;
-import lotto.application.service.LottoAnswerService;
-import lotto.application.service.LottoPurchaseService;
-import lotto.application.service.LottoResultService;
+import lotto.application.service.answer.LottoAnswerService;
+import lotto.application.service.purchase.LottoPurchaseService;
+import lotto.application.service.result.LottoResultService;
 import lotto.common.BusinessException;
 import lotto.io.input.LottoGameInputPort;
 import lotto.io.input.dto.BonusNumberRequest;
